ボタン修正
