--- src/components/CustomButtonComponent.tsx
@@ -1,32 +1,0 @@
-import React from "react";
-
-interface Props {
-  border: string;
-  color: string;
-  children?: React.ReactNode;
-  height: string;
-  onClick: () => void;
-  radius: string;
-  width: string;
-}
-
-const Button: React.FC<Props> = ({ border, color, children, height, onClick, radius, width }) => {
-  return (
-    <div>
-      <button
-        onClick={onClick}
-        style={{
-          backgroundColor: color,
-          border,
-          borderRadius: radius,
-          height,
-          width,
-        }}
-      >
-        {children}
-      </button>
-    </div>
-  );
-};
-
-export default Button;
