+++ src/components/editButton/button.tsx
@@ -1,0 +1,8 @@
+import Button from "./CustomButtonComponent";
+import React from "react";
+
+function editButton() {
+  return <Button onClick={() => console.log("clicked")} children="編集" />;
+}
+
+export default editButton;
