+++ src/components/editButton/style.module.scss
@@ -1,0 +1,7 @@
+.editButton {
+  border: 1px solid #e5e5e5;
+  color: #afafaf;
+  height: 18px;
+  border-radius: 1%;
+  max-width: 80%;
+}
